application/requirements.txt
attrs==21.4.0
click==8.1.3
flasgger==0.9.5
Flask==2.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
MarkupSafe==2.1.1
mistune==2.0.2
pyrsistent==0.18.1
PyYAML==6.0
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2